Title: 숫자 사이의 배수, Time: 139ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240903/숫자 사이의 배수/multiplication-between-numbers.java
@@ -17,7 +17,7 @@ public class Main {
                 
             
         }
-        System.out.printf("%d %.1f",sum,sum/c);
+        System.out.printf("%d %.1f",sum,(double)sum/c);
 
     }
 }
